+++ core/configs/logs.json
