+++ package.json
@@ -45,6 +45,8 @@
     "typescript": "^3.0.3"
   },
   "dependencies": {
+    "@types/morgan": "^1.7.35",
+    "@types/winston": "^2.4.4",
     "benchmark": "2.1.4",
     "express": "4.16.3",
     "fs": "0.0.1-security",
